improved spacing issues

+++ users/signals.py
@@ -13,11 +13,9 @@ def delete_profile_picture(sender, instance, **kwargs):
     Signal handler that deletes the profile picture file from the filesystem
     when a Profile instance is deleted.
     """
-    if instance.profile_picture:
+    if instance.profile_picture and os.path.isfile(instance.profile_picture.path):
 
-        if os.path.isfile(instance.profile_picture.path):
-
-            os.remove(instance.profile_picture.path)
+        os.remove(instance.profile_picture.path)
 
 @receiver(post_save, sender=CustomUser)
 def create_player_profile(sender, instance, created, **kwargs):
